feat: move onsite/reservation paths to reservation
