+++ src/apis/reservations/_id@string/check/index.ts
@@ -1,5 +1,5 @@
-import { AuthToken } from "../../../../@types";
-import { Term } from "../../../general/term/@types";
+import { AuthToken } from "../../../@types";
+import { Term } from "../../../terms/@types";
 
 export interface Methods {
   /**
